+++ src/gateway/server-methods/restart-request.ts
@@ -1,3 +1,36 @@
+/**
+ * Parse the live deliveryContext passed by gateway-tool clients.
+ *
+ * Clients capture delivery context from the active agent run and forward it
+ * so server-side handlers can write an accurate sentinel without reading the
+ * persisted session store, which heartbeat runs frequently overwrite to
+ * { channel: "webchat", to: "heartbeat" }. See #18612.
+ */
+export function parseDeliveryContextFromParams(
+  params: unknown,
+): { channel?: string; to?: string; accountId?: string } | undefined {
+  const raw = (params as { deliveryContext?: unknown }).deliveryContext;
+  if (!raw || typeof raw !== "object") {
+    return undefined;
+  }
+  const channel =
+    typeof (raw as { channel?: unknown }).channel === "string"
+      ? (raw as { channel: string }).channel.trim() || undefined
+      : undefined;
+  const to =
+    typeof (raw as { to?: unknown }).to === "string"
+      ? (raw as { to: string }).to.trim() || undefined
+      : undefined;
+  const accountId =
+    typeof (raw as { accountId?: unknown }).accountId === "string"
+      ? (raw as { accountId: string }).accountId.trim() || undefined
+      : undefined;
+  if (!channel && !to) {
+    return undefined;
+  }
+  return { channel, to, accountId };
+}
+
 export function parseRestartRequestParams(params: unknown): {
   sessionKey: string | undefined;
   note: string | undefined;
